requirements.txt
Flask==2.2.2
gunicorn==20.1.0
requests
Werkzeug==2.3.7
RPi.GPIO
